feat: Reorganize classes
