+++ DDI.DrugApi/DDI.DrugApi.csproj
@@ -8,4 +8,8 @@
 		<ProjectReference Include="../DDI.Models/DDI.Models.csproj" />
 	</ItemGroup>
 
+	<ItemGroup>
+	  <PackageReference Include="System.Net.Http.Json" Version="6.0.0" />
+	</ItemGroup>
+
 </Project>
