+++ DDI.Models/Drug.cs
@@ -1,10 +1,8 @@
-﻿using System;
-
-namespace DDI.Models
+﻿namespace DDI.Models
 {
 	public class Drug
 	{
-		public string CommonName { get; set; }
-		public string ScientificName { get; set; }
+		public int Id { get; set; }
+		public string Name { get; set; }
 	}
 }
