+++ DDI.WebApp/Startup.Dependencies.cs
@@ -1,4 +1,6 @@
-using DDI.DrugApi;
+using System.Net.Http;
+using DDI.DrugApi.Apis;
+using DDI.DrugApi.HttpClients;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace DDI.WebApp
@@ -7,6 +9,7 @@ namespace DDI.WebApp
 	{
 		public void InjectDependencies(IServiceCollection services)
 		{
+			services.AddSingleton<NlmHttpClient>();
 			services.AddScoped<IDrugApi, NlmDrugApi>();
 		}
 	}
